Security patches: RestartJob ignores pid, argv[0]; kill runs as child UID
BUG=chromium-os:7018
TEST=unit tests, install on device and verify that we can go to BWSI mode


Review URL: http://codereview.chromium.org/3493012

+++ login_manager/system_utils.cc
@@ -23,9 +23,20 @@ namespace login_manager {
 SystemUtils::SystemUtils() {}
 SystemUtils::~SystemUtils() {}
 
-int SystemUtils::kill(pid_t pid, int signal) {
-  LOG(INFO) << "Sending " << signal << " to " << pid;
-  return ::kill(pid, signal);
+int SystemUtils::kill(pid_t pid, uid_t owner, int signal) {
+  LOG(INFO) << "Sending " << signal << " to " << pid << " as " << owner;
+  uid_t uid, euid, suid;
+  getresuid(&uid, &euid, &suid);
+  if (setresuid(owner, owner, -1)) {
+    PLOG(ERROR) << "Couldn't assume uid " << owner;
+    return -1;
+  }
+  int ret = ::kill(pid, signal);
+  if (setresuid(uid, euid, -1)) {
+    PLOG(ERROR) << "Couldn't return to root";
+    return -1;
+  }
+  return ret;
 }
 
 bool SystemUtils::ChildIsGone(pid_t child_spec, int timeout) {
